chore: Move wrapper implementations
Mover the component wrappers so that they
are placed as they will be in the componens
testbench package
