+++ vaadin-testbench-core/src/test/java/com/vaadin/testbench/unit/ButtonWrapTest.java
@@ -15,11 +15,9 @@ import java.util.concurrent.atomic.AtomicReference;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import com.vaadin.flow.component.AbstractField.ComponentValueChangeEvent;
+import com.vaadin.flow.component.button.ButtonWrap;
 import com.vaadin.flow.component.ClickEvent;
-import com.vaadin.flow.component.HasValue.ValueChangeListener;
 import com.vaadin.flow.component.button.Button;
-import com.vaadin.flow.component.textfield.TextField;
 
 public class ButtonWrapTest extends UIUnitTest {
