+++ vaadin-testbench-core/src/test/java/com/vaadin/testbench/unit/WrapperResolutionTest.java
@@ -14,6 +14,7 @@ import org.junit.jupiter.api.Test;
 
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.Tag;
+import com.vaadin.flow.component.TestWrap;
 
 public class WrapperResolutionTest extends UIUnitTest {
